move utils to v17 upgrade package
