+++ app/upgrades/v17/utils_test.go
@@ -1,4 +1,4 @@
-package v16
+package v17
 
 import (
 	"fmt"
